+++ src/main/java/wdmsystem/payment/PaymentDTO.java
@@ -1,4 +1,6 @@
 package wdmsystem.payment;
 
-public record PaymentDTO(Integer id, double tipAmount, double totalAmount, PaymentMethod method, Integer orderId) {
+import java.math.BigDecimal;
+
+public record PaymentDTO(Integer id, BigDecimal tipAmount, BigDecimal totalAmount, PaymentMethod method, Integer orderId) {
 }
