+++ src/main/java/wdmsystem/product/ProductDTO.java
@@ -1,4 +1,6 @@
 package wdmsystem.product;
 
-public record ProductDTO(Integer id, Integer merchantId, String title, Integer categoryId, double price, Integer discountId, Integer taxId, float weight, String weightUnit) {
+import java.math.BigDecimal;
+
+public record ProductDTO(Integer id, Integer merchantId, String title, Integer categoryId, BigDecimal price, Integer discountId, Integer taxId, float weight, String weightUnit) {
 }
